backend/requirements.txt
# Core API dependencies
fastapi==0.103.2
uvicorn==0.22.0
python-multipart==0.0.6
aiofiles==23.1.0
requests==2.31.0

# LangChain and LangSmith - pinned versions with compatible dependencies
# Use more flexible version ranges to avoid conflicts
langchain>=0.0.354,<0.1.0
langchain-core>=0.3.48,<1.0.0
langchain-openai>=0.3.10,<0.4.0
langchain-community>=0.0.20,<0.1.0
langsmith>=0.0.77,<0.1.0

# File processing dependencies
openai>=1.3.0,<2.0.0
python-dotenv==1.0.0
# Use a binary-compatible version of pydantic to avoid Rust compilation
pydantic>=1.10.8,<2.0.0
faiss-cpu>=1.7.4
pillow>=10.0.0
numpy>=1.26.0

# OCR capabilities (optional)
paddleocr>=2.6.0.1
paddlepaddle>=2.5.0

# Audio processing
# Note: Local openai-whisper is omitted due to build issues
# Instead, using OpenAI's API for audio transcription
openai>=1.1.1  # For API-based audio transcription with Whisper

# Video processing
moviepy>=1.0.3

# Additional utilities
pyyaml>=6.0.1